Fixed SSR and SSG

+++ src/actions/queues-actions.ts
@@ -7,32 +7,50 @@ import { z } from "zod";
 import { revalidatePath } from "next/cache";
 
 export async function getQueues() {
-  const chatId = (await getSession())?.user.chatId;
-  const res = await apiClient.get("/queues", { headers: { chatId } });
+  const session = await getSession();
+  if (!session) return [];
+  const chatId = session.user.chatId;
+  if (!chatId) return [];
+  const res = await apiClient.get("/queues", {
+    headers: { chatId: chatId.toString() },
+  });
   const GetQueuesSchema = z.array(QueueSchema);
   return GetQueuesSchema.parse(res.data);
 }
 
 export async function createQueue(name: string, mixed: boolean) {
-  const chatId = (await getSession())?.user.chatId.toString();
-  await apiClient.post("/queues", { name, mixed }, { headers: { chatId } });
+  const session = await getSession();
+  if (!session) return;
+  const chatId = session.user.chatId;
+  if (!chatId) return;
+  await apiClient.post(
+    "/queues",
+    { name, mixed },
+    { headers: { chatId: chatId.toString() } },
+  );
   revalidatePath("/queues");
 }
 
 export async function shuffleQueue(queueId: string) {
-  const chatId = (await getSession())?.user.chatId.toString();
-  await apiClient.post(`/queues/shuffle/${queueId}`, { headers: { chatId } });
+  const session = await getSession();
+  if (!session) return;
+  const chatId = session.user.chatId;
+  if (!chatId) return;
+  await apiClient.post(`/queues/shuffle/${queueId}`, {
+    headers: { chatId: chatId.toString() },
+  });
   revalidatePath("/queues");
 }
 
 //TODO: Update is not working somehow
 export async function updateQueue(queue: Queue) {
   const session = await getSession();
+  if (!session) return;
   const queueDto = QueueSchema.omit({ pinned: true }).parse(queue);
   await apiClient.put("/queues", queueDto, {
     headers: {
-      chatId: session?.user.chatId.toString(),
-      userId: session?.user.accountId.toString(),
+      chatId: session.user.chatId.toString(),
+      userId: session.user.accountId.toString(),
     },
   });
 
@@ -41,9 +59,10 @@ export async function updateQueue(queue: Queue) {
 
 export async function removeQueue(queueId: string, mixed: boolean) {
   const session = await getSession();
+  if (!session) return;
   const headers = {
-    chatId: session?.user.chatId.toString(),
-    userId: session?.user.accountId.toString(),
+    chatId: session.user.chatId.toString(),
+    userId: session.user.accountId.toString(),
   };
   if (mixed) {
     await apiClient.delete(`/queues/mixed/${queueId}`, { headers });
